Conflicts:
	keystone/common/controller.py
	keystone/common/ldap/core.py
	keystone/common/models.py
	keystone/common/sql/migrate_repo/versions/009_normalize_identity_migration.py
	keystone/config.py
	keystone/identity/backends/ldap/core.py
	keystone/identity/backends/sql.py
	keystone/identity/controllers.py
	keystone/service.py
	keystone/test.py
	keystone/token/controllers.py
	tests/default_fixtures.py
	tests/test_backend_ldap.py
	tests/test_sql_upgrade.py

+++ run_tests.sh
@@ -26,6 +26,7 @@ function usage {
   echo "  -n, --no-recreate-db     Don't recreate the test database."
   echo "  -x, --stop               Stop running tests after the first error or failure."
   echo "  -f, --force              Force a clean re-build of the virtual environment. Useful when dependencies have been added."
+  echo "  -u, --update             Update the virtual environment with any newer package versions"
   echo "  -p, --pep8               Just run pep8"
   echo "  -P, --no-pep8            Don't run pep8"
   echo "  -c, --coverage           Generate coverage report"
@@ -48,7 +49,9 @@ function process_option {
     -r|--recreate-db) recreate_db=1;;
     -n|--no-recreate-db) recreate_db=0;;
     -f|--force) force=1;;
+    -u|--update) update=1;;
     -p|--pep8) just_pep8=1;;
+    -8|--8) short_pep8=1;;
     -P|--no-pep8) no_pep8=1;;
     -c|--coverage) coverage=1;;
     -xintegration) nokeystoneclient=1;;
@@ -66,13 +69,15 @@ always_venv=0
 never_venv=0
 force=0
 noseargs=
-noseopts=
+noseopts="--with-openstack --openstack-color"
 wrapper=""
 just_pep8=0
+short_pep8=0
 no_pep8=0
 coverage=0
 nokeystoneclient=0
 recreate_db=1
+update=0
 
 for arg in "$@"; do
   process_option $arg
@@ -105,6 +110,14 @@ function run_tests {
 }
 
 function run_pep8 {
+  FLAGS=--show-pep8
+  echo $#
+  if [ $# -gt 0 ] && [ 'short' == ''$1 ]
+  then
+      FLAGS=''
+  fi
+
+
   echo "Running pep8 ..."
   # Opt-out files from pep8
   ignore_scripts="*.pyc,*.pyo,*.sh,*.swp,*.rst"
@@ -113,7 +126,7 @@ function run_pep8 {
   ignore="$ignore_scripts,$ignore_files,$ignore_dirs"
   srcfiles="."
   # Just run PEP8 in current environment
-  ${wrapper} pep8 --repeat --show-source \
+  ${wrapper} pep8 --repeat $FLAGS --show-source \
     --exclude=${ignore} ${srcfiles} | tee pep8.txt
 }
 
@@ -125,6 +138,10 @@ then
   if [ $force -eq 1 ]; then
     echo "Cleaning virtualenv..."
     rm -rf ${venv}
+  fi
+  if [ $update -eq 1 ]; then
+    echo "Updating virtualenv..."
+    python tools/install_venv.py
   fi
   if [ -e ${venv} ]; then
     wrapper="${with_venv}"
@@ -154,6 +171,12 @@ if [ $just_pep8 -eq 1 ]; then
     run_pep8
     exit
 fi
+
+if [ $short_pep8 -eq 1 ]; then
+     run_pep8 short
+     exit
+fi
+
 
 if [ $recreate_db -eq 1 ]; then
     rm -f tests.sqlite
